requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
execnet==2.1.2
Faker==40.12.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
mailslurp-client==17.5.0
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.20.0
PySocks==1.7.1
pytest==9.0.2
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
